List Unit Test: All list_spec.rb test are passing.
All List unit tests in list_spec.rb are now passing using the new refactor. Only two factories are used at the top, with more used to fill the lists as needed in test. Redundant code and tests removed, and all behavior previously tested remains, with added test of testing if a List is valid (which it is if and only if it has 5 champions). References #5.