Workflow Tests: Update creates_list_spec to attach 5 champions to satisfy list length enforcement.
Enforcing champions list length with custom ActiveRecord validator was not allowing the creates_list_spec 'attaches champions to the list' test to successfully build the list and not be a new record. With a test list of 5 champions parsed into the new list we can expect the list size to be 5 and the record to not be new since it adds them to the 'creator' instantiated on line 4. References #6.