Replace Champion: Create replace method to destroy an existing champion and replace with new.
Refactor list spec to isolate db dependency (add a test object via create specifically for the replace method unit test). Preserved the existing test objects, but under more descriptive names. The 'is not valid if size is not five' test is not complete and certainly not a good test! Ben and Brian team struggle bus commit. References #4