Replace Champion: Refactor teams for testing and use create method for 'replace' unit test.
Improved naming for list_spec unit tests. Created new team for testing the replace method specifically, which utilizes the 'create' method. This instantiates the objects with IDs and allows us to leverage the objects themselves to test the method. In order for this to work, the champion class had to be given the belongs to, 'optional' characterization. The unit test force reloads the list to ensure that the deletion of the old champion is registered. References #4.